Fixing error in CSR (by disabling challenge phrase) when creating using ispConfig:
https://ssltools.websecurity.symantec.com/checker/views/csrCheck.jsp
You have 1 error
Your CSR contains a challenge phrase.
Adding a challenge phrase to a CSR is not a secure practice. Please generate a new CSR that does not contain a challenge phrase.

The problem occure while buying a cert from https://www.ssls.com/
The support reply with:
"Please regenerate it without the passphrase and try to activate the certificate once again."