Warn if the lockfile for system cron is older than a day
We want to avoid deadlock situations and right now if there is a process
running with the same PID that is not our process (e.g. after a reboot)
we might be waiting forever to run the cron again. So to make this
transparent to the user, we at least post a warning into the logfile.